Extract classes into files and bump redscript for an include guard
